Refactor

+++ sha2/src/lib.rs
@@ -20,9 +20,8 @@ cpufeatures::new!(cpuid_sha, "sha");
 #[allow(dead_code)]
 extern "C" {
     fn sha256_compress(state: &mut [u32; 8], block: &[u8; 64]);
-    // setting usize is safe, cause we are on 64 bit platform
-    fn sha256_transform_rorx(state: &mut [u32; 8], block: *const [u8; 64], num_blocks: usize);
-    fn sha256_ni_transform(digest: &mut [u32; 8], data: *const [u8; 64], nblk: usize);
+    fn sha256_transform_rorx(state: &mut [u32; 8], block: *const [u8; 64], num_blocks: u64);
+    fn sha256_ni_transform(digest: &mut [u32; 8], data: *const [u8; 64], nblk: u64);
 }
 
 /// Safe wrapper around assembly implementation of SHA256 compression function
@@ -32,7 +31,7 @@ pub fn compress256(state: &mut [u32; 8], blocks: &[[u8; 64]]) {
     let token_sha = cpuid_sha::init();
     if token_sha.get() {
         if !blocks.is_empty() {
-            unsafe { sha256_ni_transform(state, blocks.as_ptr(), blocks.len()) }
+            unsafe { sha256_ni_transform(state, blocks.as_ptr(), blocks.len() as u64) }
         }
         return;
     }
@@ -40,7 +39,7 @@ pub fn compress256(state: &mut [u32; 8], blocks: &[[u8; 64]]) {
 
     if token.get() {
         if !blocks.is_empty() {
-            unsafe { sha256_transform_rorx(state, blocks.as_ptr(), blocks.len()) }
+            unsafe { sha256_transform_rorx(state, blocks.as_ptr(), blocks.len() as u64) }
         }
     } else {
         for block in blocks {
@@ -53,7 +52,7 @@ pub fn compress256(state: &mut [u32; 8], blocks: &[[u8; 64]]) {
 #[link(name = "sha512", kind = "static")]
 extern "C" {
     fn sha512_compress(state: &mut [u64; 8], block: &[u8; 128]);
-    fn sha512_transform_rorx(state: &mut [u64; 8], block: *const [u8; 128], num_blocks: usize);
+    fn sha512_transform_rorx(state: &mut [u64; 8], block: *const [u8; 128], num_blocks: u64);
 }
 
 /// Safe wrapper around assembly implementation of SHA512 compression function
@@ -65,7 +64,7 @@ pub fn compress512(state: &mut [u64; 8], blocks: &[[u8; 128]]) {
     let token: cpuid_avx2::InitToken = cpuid_avx2::init();
     if token.get() {
         if !blocks.is_empty() {
-            unsafe { sha512_transform_rorx(state, blocks.as_ptr(), blocks.len()) }
+            unsafe { sha512_transform_rorx(state, blocks.as_ptr(), blocks.len() as u64) }
         }
     } else {
         for block in blocks {
